Add files via upload
This project demonstrates how to use LabVIEW in conjunction with an Arduino microcontroller to monitor light intensity via an LDR (Light Dependent Resistor) sensor. The setup involves reading the LDR sensor values using the Arduino and sending this data to LabVIEW for real-time processing and control of LEDs based on the measured light intensity.

The primary goal of this project is to:

Read light intensity values from an LDR sensor connected to an Arduino.
Process these values in LabVIEW.
Control two LEDs of LabVIEW based on the light intensity:
When the light intensity is less than 40, LED2 will turn off.
When the light intensity is 40 or greater, LED1 will turn on.

Happy tinkering!